Updated video embed.
It turns out that Vimeo allows only Plus members to convert a video to a mobile-friendly format. See: http://vimeo.com/faq#mobile

Tried YouTube which mostly works. But am noticing that the video image changes when on an iPad/iPhone compared to on a desktop browser. It actually letterboxes it even though it doesn't need to.

I could conceivably convert our video to an iPad/iPhone-friendly QuickTime movie and we can also upload to YouTube and Vimeo for social sharing.