Retinize the screenshot images.
Done by using the somewhat hinky `retina.js` API. Once that worked, realized
that I needed to set `innerHeight` and `innerWidth` for the colorbox so that
they were scaled properly. Looks pretty good now; only the images from
`colorbox.css` are not retinized. Not sure how important that is. Issue #10.